Convert TestServer to pipes. #11598
