+++ src/Hosting/TestHost/src/ClientHandler.cs
@@ -64,7 +64,6 @@ namespace Microsoft.AspNetCore.TestHost
 
             var contextBuilder = new HttpContextBuilder(_application, AllowSynchronousIO, PreserveExecutionContext);
 
-            Stream responseBody = null;
             var requestContent = request.Content ?? new StreamContent(Stream.Null);
             var body = await requestContent.ReadAsStreamAsync();
             contextBuilder.Configure(context =>
@@ -114,8 +113,6 @@ namespace Microsoft.AspNetCore.TestHost
                     body.Seek(0, SeekOrigin.Begin);
                 }
                 req.Body = new AsyncStreamWrapper(body, () => contextBuilder.AllowSynchronousIO);
-
-                responseBody = context.Response.Body;
             });
 
             var response = new HttpResponseMessage();
@@ -138,7 +135,7 @@ namespace Microsoft.AspNetCore.TestHost
             response.ReasonPhrase = httpContext.Features.Get<IHttpResponseFeature>().ReasonPhrase;
             response.RequestMessage = request;
 
-            response.Content = new StreamContent(responseBody);
+            response.Content = new StreamContent(httpContext.Response.Body);
 
             foreach (var header in httpContext.Response.Headers)
             {
